More docs updates, and a fix to currently non-functioning wv_relay code
