+++ lacci/lib/lacci/version.rb
@@ -1,5 +1,13 @@
 # frozen_string_literal: true
 
+# The name "scarpe" means "shoes" in Italian. "Lacci"
+# means "laces." Lacci is the display-independent part
+# of Scarpe, tying the Shoes API to the display service
+# of your choice.
+#
+# Almost nothing is called Lacci in your program. It's
+# mostly invisible. Instead, look at the {Shoes} module
+# to see what's in Lacci.
 module Lacci
   VERSION = "0.2.1"
 end
